Melhorias/Novas funcionalidades
Adicionado as funcionalidades de registro, criado logica de acesso com retorno de tokem para o usuário

+++ apiS/apiBanco/api/src/graphql/resolvers/apiResolversUsuarios.js
@@ -1,15 +1,41 @@
 const usuarioModel = require('../../models/usuario');
+const jwt = require('jwt-simple');
+const moment = require('moment');
+const segredo = 'sdmiuuebf375reuwd4SDjsdJAHs'
 
 const apiResolversUsuario = {
 
-  getUsuarios: () => {
-    let usuario = usuarioModel.find({},{Acesso:0, Configuracao:0}).exec()//function (err, docs) {console.log(docs)}
+  login: async ({input}) => {
 
-    if (!usuario) {
-      return {cdRetorno: -2, dsRetorno: 'Erro ao buscar os usuários do Banco de Dados'}
+    let usuario = await usuarioModel.findOne({Nome: input.Nome},{Acessos:0, Configuracao:0}).exec()//function (err, docs) {console.log(docs)}
+
+    var expires = moment().add(1,'days').valueOf();
+    if(usuario.Senha == input.Senha) {
+        var token = jwt.encode({
+          iss: usuario._id,
+          exp: expires
+        }, segredo);
+
+        return {cdRetorno: 0, dsRetorno: 'Sucesso', token: token}
+    } else {
+        return {cdRetorno: -1, dsRetorno: 'Usuario ou senha invalidos!'}
     }
 
-    return usuario
+  },
+
+  getUsuarios: async ({input}) => {
+
+    let validToken = await verificaToken(input.token)
+
+    if(!validToken)
+      return {cdRetorno: -1, dsRetorno: 'Token não informado, invalido ou expirado!'}
+
+    let usuarios = await usuarioModel.find({},{Acessos:0, Configuracao:0}).exec()//function (err, docs) {console.log(docs)}
+
+    if (!usuarios)
+      return {cdRetorno: -2, dsRetorno: 'Erro ao buscar os usuários do Banco de Dados'}
+
+    return {cdRetorno: 0, dsRetorno: 'Sucesso', Usuarios: usuarios}
   },
 
   createUsuario: ({input}) => {
@@ -17,8 +43,9 @@ const apiResolversUsuario = {
     create = {}
     try {
       create["Nome"]=input.Nome;
+      create["Senha"]=input.Senha;
       create["IPaddres"]=input.IPaddres;
-      create["Configuracao"]={usuario: "Teste"};
+      create["Configuracao"]={usuario: input.Nome};
 
       let uModel = new usuarioModel(create);
       let newUsuarioInDB = uModel.save();
@@ -32,6 +59,11 @@ const apiResolversUsuario = {
   },
 
   updateUsuario: async ({input}) => {
+
+    let validToken = await verificaToken(input.token)
+
+    if(!validToken)
+      return {cdRetorno: -1, dsRetorno: 'Token não informado, invalido ou expirado!'}
 
     var update={};
 
@@ -48,9 +80,9 @@ const apiResolversUsuario = {
         { upsert: false }
       ).exec()
 
-      if (!newUpdate) {
+      if (!newUpdate)
         return {cdRetorno: -1, dsRetorno: 'Usuário não existe!'}
-      }
+
       console.log("Usuário atualizado: "+newUpdate)
 
       let atualizacoes = usuarioModel.update(
@@ -69,13 +101,18 @@ const apiResolversUsuario = {
 
   deleteUsuario: async ({input}) => {
 
+    let validToken = await verificaToken(input.token)
+
+    if(!validToken)
+      return {cdRetorno: -1, dsRetorno: 'Token não informado, invalido ou expirado!'}
+
     try {
 
       let removeUsuario = await usuarioModel.findOneAndRemove({ _id:input._id }).exec()
 
-      if (!removeUsuario) {
+      if (!removeUsuario)
         return {cdRetorno: -1, dsRetorno: 'Usuário não existe!'}
-      }
+
       console.log("Usuário removido: "+removeUsuario)
       return {cdRetorno: 0, dsRetorno: 'Sucesso'}
 
@@ -83,8 +120,57 @@ const apiResolversUsuario = {
         return {cdRetorno: -2, dsRetorno: 'Erro ao remover o usuário do Banco de Dados, erro: '+e}
     }
 
+  },
+
+  registraAcesso: async ({input}) => {
+    let validToken = await verificaToken(input.token)
+
+    if(!validToken)
+      return {cdRetorno: -1, dsRetorno: 'Token não informado, invalido ou expirado!'}
+
+  },
+
+  registraPesquisa: async ({input}) => {
+    let validToken = await verificaToken(input.token)
+
+    if(!validToken)
+      return {cdRetorno: -1, dsRetorno: 'Token não informado, invalido ou expirado!'}
+
+  },
+
+  registraChat: async ({input}) => {
+    let validToken = await verificaToken(input.token)
+
+    if(!validToken)
+      return {cdRetorno: -1, dsRetorno: 'Token não informado, invalido ou expirado!'}
+      
   }
 
 };
+
+async function verificaToken(token) {
+  try {
+    var decoded = jwt.decode(token, segredo);
+    console.log('decodando ' + decoded.iss);
+
+    if (decoded.exp <= Date.now()) {
+      console.log('Data expirada: '+decoded.exp);
+      return false
+    }
+
+    let usuario = await usuarioModel.findOne({ _id: decoded.iss }).exec()
+
+    if(!usuario) {
+      console.log("Usuário não encontrado com o id: "+decoded.iss);
+      return false
+    } else {
+      return true
+    }
+
+  } catch (e) {
+    console.log('Exception: '+e);
+    return false
+  }
+}
 
 module.exports = apiResolversUsuario
